Final_FakeNews/requirements.txt
streamlit
transformers
torch
requests
python-dotenv
huggingface_hub
